Do not process list of dicts as expected data for role consumption
Rather than expecting user to pass a list of dicts containing
the src image path and the image name, just leverage the include_role
with_items to do so.
That way the role logic stays the same, irrespective of user passing
a single pair of src image path and image name or a list.
The README contains an example on how to do so in case of a list.